Update dependencies on acitvation task and organizue input variables
