+++ Tasks/UnityActivateLicense/UnityActivateLicenseV1/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2]
+
+### Changed
+
+- Aligned configuraiton options grouping and naming with other tasks
+- Updated and cleaned up external dependencies
+
 ## [1.2.1]
 
 ### Changed
