+++ Tasks/UnityActivateLicense/UnityActivateLicenseV1/task.json
@@ -12,7 +12,7 @@
   "version": {
     "Major": "1",
     "Minor": "2",
-    "Patch": "1"
+    "Patch": "2"
   },
   "releaseNotes": "[Full Changelog](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/blob/master/Tasks/UnityActivateLicense/UnityActivateLicenseV1/CHANGELOG.md)",
   "minimumAgentVersion": "2.144.0",
@@ -50,6 +50,14 @@
       "helpMarkDown": "Serial key used to activate the Unity Editor. The serial key must be assigned to the user defined by username and password above."
     },
     {
+      "name": "deactivateSeatOnComplete",
+      "type": "boolean",
+      "label": "Deactivate seat on complete",
+      "defaultValue": true,
+      "required": false,
+      "helpMarkDown": "By default the seat activation is released again after the pipeline has run. Uncheck this to keep the activation active."
+    },
+    {
       "name": "versionSelectionMode",
       "type": "pickList",
       "label": "Unity version to use activate the license with",
@@ -62,6 +70,15 @@
       }
     },
     {
+      "name": "unityProjectPath",
+      "type": "filePath",
+      "label": "Unity project path",
+      "defaultValue": "",
+      "required": false,
+      "helpMarkDown": "(Optional) Enter the path to the Unity project within the repository. If no value is entered, the root of the repository will be used.",
+      "visibleRule": "versionSelectionMode == project"
+    },
+    {
       "name": "version",
       "type": "string",
       "label": "Version",
@@ -71,42 +88,24 @@
       "visibleRule": "versionSelectionMode == specify"
     },
     {
-      "name": "unityEditorsPathMode",
+      "name": "unityEditorsLocation",
       "type": "pickList",
       "label": "Unity editors location",
-      "defaultValue": "unityHub",
+      "defaultValue": "default",
       "helpMarkDown": "Define where to look for Unity installations on the agent.",
       "options": {
-        "unityHub": "Unity Hub default installation path",
-        "environmentVariable": "Use UNITYHUB_EDITORS_FOLDER_LOCATION environment variable on agent",
+        "default": "Default (recommended)",
         "specify": "Specify path"
       }
     },
     {
-      "name": "customUnityEditorsPath",
+      "name": "customUnityEditorsLocation",
       "type": "string",
       "label": "Editors folder location",
       "defaultValue": "",
       "required": true,
-      "visibleRule": "unityEditorsPathMode == specify",
+      "visibleRule": "unityEditorsLocation == specify",
       "helpMarkDown": "Specify where to look for Unity Editor versions on the agent."
-    },
-    {
-      "name": "unityProjectPath",
-      "type": "filePath",
-      "label": "Unity project path",
-      "defaultValue": "",
-      "required": false,
-      "helpMarkDown": "(Optional) Enter the directory path to the Unity project. If no value is entered, the root of the repository will be used."
-    },
-    {
-      "name": "deactivateSeatOnComplete",
-      "type": "boolean",
-      "label": "Deactivate seat on complete",
-      "defaultValue": true,
-      "required": true,
-      "groupName": "advanced",
-      "helpMarkDown": "By default the seat activation is released again after the pipeline has run. Uncheck this to keep the activation active."
     }
   ],
   "outputVariables": [
@@ -116,13 +115,13 @@
     }
   ],
   "execution": {
-    "Node10": {
-      "target": "unity-activate-license.js",
+    "Node": {
+      "target": "unity-setup.js",
       "argumentFormat": ""
     }
   },
   "postjobexecution": {
-    "Node10": {
+    "Node": {
       "target": "post-unity-activate-license.js",
       "argumentFormat": ""
     }
