+++ Tasks/UnitySetup/UnitySetupV1/unity-setup.ts
@@ -1,6 +1,7 @@
 import path = require('path');
 import tl = require('azure-pipelines-task-lib/task');
-import { OS, UnityPathTools, UnityVersionInfoResult, UnityVersionTools, Utilities } from '@dinomite-studios/unity-azure-pipelines-tasks-lib/';
+import { OS, UnityPathTools, Utilities } from '@dinomite-studios/unity-azure-pipelines-tasks-lib/';
+import { getUnityEditorVersion } from './unity-build-shared';
 
 // Input variables - General.
 const versionSelectionModeVariableName = "versionSelectionMode";
@@ -175,27 +176,6 @@ function run() {
             tl.setResult(tl.TaskResult.Failed, `${e}`);
         }
     }
-}
-
-function getUnityEditorVersion(): UnityVersionInfoResult {
-    const projectPath = tl.getPathInput('unityProjectPath') || '';
-    console.log(`${tl.loc('projectPathInfo')} ${projectPath}`);
-
-    const unityVersion = UnityVersionTools.determineProjectVersionFromFile(projectPath);
-    if (unityVersion.error) {
-        const error = `${tl.loc('failGetUnityEditorVersion')} | ${unityVersion.error}`;
-        console.error(error);
-        throw new Error(error);
-    }
-
-    const successGetVersionLog = `${tl.loc('successGetUnityEditorVersion')} ${unityVersion.info!.version}${unityVersion.info!.revision ? `, revision=${unityVersion.info!.revision}` : ''}, alpha=${unityVersion.info!.isAlpha}, beta=${unityVersion.info!.isBeta}`;
-    console.log(successGetVersionLog);
-
-    if (unityVersion.info!.isAlpha || unityVersion.info!.isBeta) {
-        console.warn(tl.loc('warningAlphaBetaVersion'));
-    }
-
-    return unityVersion;
 }
 
 run();
